Fix Assertion Const Node Tensor Request when Node has value_int Rather than Value (#181)
This PR adds a check to extract_constant_to_initializer and the
FetchConstantTensor function (in pass_util.h) that ensures the constant
node actually contains a Tensor (rather than a value_int or some other
type). If not checked, this causes an assertion failure when these pass
and function are used.

See Issue: https://github.com/onnx/optimizer/issues/173

Signed-off-by: Joshua Monson <joshmonson@microsoft.com>
Co-authored-by: Joshua Monson <joshmonson@microsoft.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>